% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_file_data}
\alias{extract_file_data}
\title{Extract xml data from a file on disk}
\usage{
extract_file_data(xml_file, quiet = FALSE)
}
\arguments{
\item{xml_file}{xml containing data (fetched from OSM).}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\value{
A list containing the following elements :
\itemize{
\item \code{bounds} : the bounding box of the data ;
\item \code{rel_tags} : the tags of the relation in the OpenStreetMap database ;
\item \code{rel_attr} : metadata of the relation, such as id, version, date of
last modification... ;
\item \code{stop_count} : number of stops ;
\item \code{trkpt_count}: numbre of points of the track ;
\item \code{trk_km} : length of the track, in kilometers ;
\item \code{stop_base} : data.frame with informations about stops ;
\item \code{trkpt_base} : data.frame with informations about trackpoints (in
particular the distance between too consecutive points, in meters)
}
}
\description{
Extract xml data from a file on disk
}
\keyword{internal}
