% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpx.R
\name{write_gpx}
\alias{write_gpx}
\title{Export bus line to GPX}
\usage{
write_gpx(
  id_rel,
  path = ".",
  filename = NULL,
  osm_info = TRUE,
  overpass_url = "http://overpass-api.de/api/interpreter",
  quiet = FALSE
)
}
\arguments{
\item{id_rel}{the identifier of the OpenStreetMap relation.}

\item{path}{the directory where to write the GPX file.}

\item{filename}{name of the file. If \code{NULL} (default), a name based on the
tags \code{route}, \code{ref} and \code{id} of the OSM relation.}

\item{osm_info}{include informations about OpenStreetMap nodes (id and
version).}

\item{overpass_url}{instance of the Overpass API to use to retrieve data.}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\value{
The data used to export the file, invisibly. This is the same as the
result of \code{\link{extract_data}}.
}
\description{
Export bus line to GPX
}
