% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gpx.R
\name{write_gpx_from_list}
\alias{write_gpx_from_list}
\title{Export the list of information about a transport line to GPX}
\usage{
write_gpx_from_list(
  data_list,
  path = ".",
  filename = NULL,
  osm_info = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data_list}{list of informations, typically given by \code{\link{extract_data}}.}

\item{path}{the directory where to write the GPX file.}

\item{filename}{name of the file. If \code{NULL} (default), a name based on the
tags \code{route}, \code{ref} and \code{id} of the OSM relation.}

\item{osm_info}{include informations about OpenStreetMap nodes (id and
version).}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\value{
The data used to export the file, invisibly. This is the same as the
result of \code{\link{extract_data}}.
}
\description{
Export the list of information about a transport line to GPX
}
\section{Content of the file}{

\itemize{
\item transport line metadata : some retrieved from OSM (for example the version
of the relation) and some calculated by the package (for example the length
in kilometers) ;
\item coordinates and names of the stops (role = "stop" in the relation), as
\code{wpt(s)} elements ;
\item list of the track points in the correct order, as a \code{trkseg} element.
}
}

\keyword{internal}
