% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Download data and procces it to be exported}
\usage{
extract_data(
  id_rel,
  overpass_url = "http://overpass-api.de/api/interpreter",
  quiet = FALSE
)
}
\arguments{
\item{id_rel}{the identifier of the OpenStreetMap relation.}

\item{overpass_url}{instance of the Overpass API to use to retrieve data.}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\description{
Download data and procces it to be exported
}
