% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_export_data.R
\name{write_gpx}
\alias{write_gpx}
\title{Export bus line to GPX}
\usage{
write_gpx(
  id_rel,
  path = ".",
  overpass_url = "http://overpass-api.de/api/interpreter",
  quiet = FALSE
)
}
\arguments{
\item{id_rel}{the identifier of the OpenStreetMap relation.}

\item{path}{the directorty where to write the GPX file.}

\item{overpass_url}{instance of the Overpass API to use to retrieve data.}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\value{
The data used to export the file, invisibly. This is the same as the
result of \code{\link{extract_data}}.
}
\description{
Export bus line to GPX
}
