% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_export_data.R
\name{write_gpx}
\alias{write_gpx}
\title{Export bus line to GPX}
\usage{
write_gpx(
  id_rel,
  path = ".",
  overpass_url = getOption("osmbus.overpass_url"),
  quiet = FALSE
)
}
\arguments{
\item{id_rel}{the identifier of the OpenStreetMap relation.}

\item{path}{the directorty where to write the GPX file.}

\item{overpass_url}{instance of the Overpass API to use to retrieve data.}

\item{quiet}{\code{TRUE} to desactivate information messages.}
}
\description{
Export bus line to GPX
}
